% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dm_plotData.R
\name{dm_plotDataDSInfo}
\alias{dm_plotDataDSInfo}
\title{Plot the frequency of present features}
\usage{
dm_plotDataDSInfo(info, ds_info)
}
\arguments{
\item{info}{Data frame with \code{gene_id} and \code{feature_id} of ALL
features}

\item{ds_info}{Data frame with \code{gene_id} and \code{feature_id} of ONLY
  DS features}
}
\value{
\code{ggplot} object
}
\description{
Plot the frequency of present features
}

