% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_twoStageTest.R
\name{dm_twoStageTest}
\alias{dm_twoStageTest}
\title{Perform the two stage test}
\usage{
dm_twoStageTest(pvalue_gene, pvalue_feature, FDR = 0.05, verbose = FALSE)
}
\arguments{
\item{pvalue_gene}{data frame with pvalue and gene_id}

\item{pvalue_feature}{data frame with pvalue, gene_id and feature_id}

\item{FDR}{numeric cutoff for the FDR}

\item{verbose}{logical}
}
\value{
Returns a data frame with adjusted feature level p-values
}
\description{
Perform the two stage test
}

