% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmDS_two_stage_test.R
\name{dmDS_two_stage_test}
\alias{dmDS_two_stage_test}
\title{Perform the two stage test}
\usage{
dmDS_two_stage_test(pvalue_gene, pvalue_feature, FDR = 0.05,
  verbose = FALSE)
}
\arguments{
\item{pvalue_gene}{data frame with pvalue and gene_id}

\item{pvalue_feature}{data frame with pvalue, gene_id and feature_id}

\item{FDR}{numeric cutoff for the FDR}

\item{verbose}{logical}
}
\value{
Returns a data frame with adjusted feature level p-values
}
\description{
Perform the two stage test
}

