% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\name{dmDSdata}
\alias{dmDSdata}
\title{Create dmDSdata object}
\usage{
dmDSdata(counts, samples)
}
\arguments{
\item{counts}{Data frame with counts. Rows correspond to features, for
example, transcripts or exons. This data frame has to contain a
\code{gene_id} column with gene IDs, \code{feature_id} column with feature
IDs and columns with counts for each sample. Column names corresponding to
sample IDs must be the same as in the \code{sample} data frame.}

\item{samples}{Data frame where each row corresponds to one sample. Columns
have to contain unique sample IDs in \code{sample_id} variable and a
grouping variable \code{group}.}
}
\value{
Returns a \linkS4class{dmDSdata} object.
}
\description{
Constructor function for a \code{\linkS4class{dmDSdata}} object.
}
\examples{
###################################
### Differential splicing analysis
###################################
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{plotData}}, \code{\link{dmFilter}}, 
  \code{\link{dmDispersion}}, \code{\link{dmFit}}, \code{\link{dmTest}}
}

