% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSfit.R
\docType{class}
\name{dmDSfit-class}
\alias{dmDSfit-class}
\alias{proportions}
\alias{proportions,dmDSfit-method}
\alias{statistics}
\alias{statistics,dmDSfit-method}
\title{dmDSfit object}
\usage{
proportions(x, ...)

\S4method{proportions}{dmDSfit}(x)

statistics(x, ...)

\S4method{statistics}{dmDSfit}(x)
}
\arguments{
\item{x}{dmDSdispersion object.}

\item{...}{Other parameters that can be defined by methods using this
generic.}
}
\value{
\itemize{ \item \code{proportions(x)}: Get a data frame with estimated
feature ratios for each condition. \item \code{statistics(x)}: Get a data
frame with maximum log-likelihoods for each condition. }
}
\description{
dmDSfit extends the \code{\linkS4class{dmDSdispersion}} class by adding the
full model Dirichlet-multinomial feature proportion estimates needed for the
differential splicing analysis. Feature ratios are estimated for each gene
and each condition. Result of \code{\link{dmFit}}.
}
\section{Slots}{

\describe{
\item{\code{dispersion}}{Character specifying which type of dispersion was used for
fitting: \code{"common_dispersion"} or \code{"genewise_dispersion"}.}

\item{\code{fit_full}}{\code{\linkS4class{MatrixList}} containing the per gene
feature ratios. Columns correspond to different conditions. Additionally,
the full model likelihoods are stored in \code{metadata} slot.}
}}
\examples{
###################################
### Differential splicing analysis
###################################
# If possible, use BPPARAM = BiocParallel::MulticoreParam() with more workers

d <- data_dmDSdata
\donttest{
### Filtering
# Check what is the minimal number of replicates per condition 
table(samples(d)$group)
d <- dmFilter(d, min_samps_gene_expr = 7, min_samps_feature_expr = 3, 
 min_samps_feature_prop = 0)

### Calculate dispersion
d <- dmDispersion(d, BPPARAM = BiocParallel::SerialParam())

### Fit full model proportions
d <- dmFit(d, BPPARAM = BiocParallel::SrialParam())

head(proportions(d))
head(statistics(d))

}

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\linkS4class{dmDSdata}},
  \code{\linkS4class{dmDSdispersion}}, \code{\linkS4class{dmDStest}}
}

