% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDStest.R
\docType{class}
\name{dmDStest-class}
\alias{dmDStest-class}
\alias{results}
\alias{results,dmDStest-method}
\title{dmDStest object}
\usage{
results(x, ...)

\S4method{results}{dmDStest}(x, level = "gene")
}
\arguments{
\item{x}{dmDStest object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}

\item{level}{Character specifying which type of results to return. Possible
values \code{"gene"} or \code{"feature"}.}
}
\value{
\itemize{ \item \code{results(x)}: Get a data frame with gene-level or
feature-level results. See Slots. }
}
\description{
dmDStest extends the \code{\linkS4class{dmDSfit}} class by adding the null 
model Dirichlet-multinomial feature proportion estimates and the results of 
testing for differential splicing. Proportions are calculated for each gene 
from pooled (no grouping into conditions) counts. Result of 
\code{\link{dmTest}}.
}
\section{Slots}{

\describe{
\item{\code{design_fit_null}}{Numeric matrix of the desing used to fit the null 
model.}

\item{\code{lik_null}}{Numeric vector of the per gene DM null model likelihoods.}

\item{\code{lik_null_bb}}{Numeric vector of the per gene BB null model likelihoods.}

\item{\code{results_gene}}{Data frame with the gene-level results including: 
\code{gene_id} - gene IDs, \code{lr} - likelihood ratio statistics based on
the DM model, \code{df} - degrees of freedom, \code{pvalue} - p-values and 
\code{adj_pvalue} - Benjamini & Hochberg adjusted p-values.}

\item{\code{results_feature}}{Data frame with the feature-level results including: 
\code{gene_id} - gene IDs, \code{feature_id} - feature IDs, \code{lr} - 
likelihood ratio statistics based on the BB model, \code{df} - degrees of 
freedom, \code{pvalue} - p-values and \code{adj_pvalue} - Benjamini & 
Hochberg adjusted p-values.}
}}
\examples{

###################################
### Differential splicing analysis
###################################

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\linkS4class{dmDSdata}}, 
  \code{\linkS4class{dmDSdispersion}}, \code{\linkS4class{dmDSfit}}
}

