% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSfit.R, R/class_dmSQTLfit.R
\docType{methods}
\name{dmFit}
\alias{dmFit}
\alias{dmFit,dmDSdispersion-method}
\alias{dmFit,dmSQTLdispersion-method}
\title{Estimate proportions in Dirichlet-multinomial model}
\usage{
dmFit(x, ...)

\S4method{dmFit}{dmDSdispersion}(x, design, prop_mode = "constrOptimG",
  prop_tol = 1e-12, verbose = 0, BPPARAM = BiocParallel::SerialParam())

\S4method{dmFit}{dmSQTLdispersion}(x, dispersion = "genewise_dispersion",
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = 0,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdispersion}} or
\code{\linkS4class{dmSQTLdispersion}} object.}

\item{...}{Other parameters that can be defined by methods using this
generic.}

\item{design}{Numeric matrix definig the full model.}

\item{prop_mode}{Optimization method used to estimate proportions. Possible 
values \code{"constrOptim"} and \code{"constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating proportions.}

\item{verbose}{Numeric. Definie the level of progress messages displayed. 0 -
no messages, 1 - main messages, 2 - message for every gene fitting.}

\item{BPPARAM}{Parallelization method used by 
\code{\link[BiocParallel]{bplapply}}.}
}
\value{
Returns a \code{\linkS4class{dmDSfit}} or 
  \code{\linkS4class{dmSQTLfit}} object.
}
\description{
Maximum likelihood estimates of genomic feature (for instance, transcript,
exon, exonic bin) proportions in full Dirichlet-multinomial model used in
differential splicing or sQTL analysis. Full model estimation means that
proportions are estimated for every group/condition separately.
}
\examples{
###################################
### Differential splicing analysis
###################################

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}}, 
  \code{\link{plotFit}}, \code{\link{dmDispersion}}, \code{\link{dmTest}}
}

