% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmSQTLdispersion.R
\docType{class}
\name{dmSQTLdispersion-class}
\alias{dmSQTLdispersion-class}
\title{dmSQTLdispersion object}
\description{
dmSQTLdispersion extends the \code{\linkS4class{dmSQTLdata}} by adding the
dispersion estimates of Dirichlet-multinomial distribution used to model the
feature (e.g., transcript, exon, exonic bin) counts for each gene-SNP pair in
the sQTL analysis. Result of \code{\link{dmDispersion}}.
}
\section{Slots}{

\describe{
\item{\code{mean_expression}}{Numeric vector of mean gene expression.}

\item{\code{common_dispersion}}{Numeric value of estimated common dispersion.}

\item{\code{genewise_dispersion}}{List of estimated gene-wise dispersions. Each
element of this list is a vector of dispersions estimated for all the
genotype blocks assigned to a given gene.}
}}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmSQTLdata}}, \code{\linkS4class{dmSQTLdata}},
  \code{\linkS4class{dmSQTLfit}}, \code{\linkS4class{dmSQTLtest}}
}

