% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{dmTest}
\alias{dmTest}
\alias{dmTest,dmDSfit-method}
\alias{dmTest,dmSQTLfit-method}
\title{Likelihood ratio test}
\usage{
dmTest(x, ...)

\S4method{dmTest}{dmDSfit}(x, coef = NULL, design = NULL, contrast = NULL,
  one_way = TRUE, bb_model = TRUE, prop_mode = "constrOptim",
  prop_tol = 1e-12, coef_mode = "optim", coef_tol = 1e-12, verbose = 0,
  BPPARAM = BiocParallel::SerialParam())

\S4method{dmTest}{dmSQTLfit}(x, permutations = "all_genes",
  prop_mode = "constrOptim", prop_tol = 1e-12, verbose = 0,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}} 
object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}

\item{coef}{Integer or character vector indicating which coefficients of the 
linear model are to be tested equal to zero. Values must indicate column 
numbers or column names of the \code{design} used in 
\code{\linkS4class{dmFit}}.}

\item{design}{Numeric matrix definig the null model.}

\item{contrast}{Numeric vector or matrix specifying one or more contrasts of 
the linear model coefficients to be tested equal to zero. Number of rows 
must equal to the number of columns of \code{design} used in 
\code{\linkS4class{dmFit}}.}

\item{one_way}{Logical. Should the shortcut fitting be used when the design 
corresponds to multiple group comparison. If \code{TRUE} (the default), 
then proportions are fitted per group and regression coefficients are 
recalculated from these fittings.}

\item{bb_model}{Logical. Whether to perform the feature-level analysis using
the beta-binomial model.}

\item{prop_mode}{Optimization method used to estimate proportions. Possible 
value \code{"constrOptim"}.}

\item{prop_tol}{The desired accuracy when estimating proportions.}

\item{coef_mode}{Optimization method used to estimate regression
coefficients. Possible values \code{"optim"} (the default), \code{"nlminb"}
or \code{"Rcgmin"}.}

\item{coef_tol}{The desired accuracy when estimating regression coefficients.}

\item{verbose}{Numeric. Definie the level of progress messages displayed. 0 -
no messages, 1 - main messages, 2 - message for every gene fitting.}

\item{BPPARAM}{Parallelization method used by 
\code{\link[BiocParallel]{bplapply}}.}

\item{permutations}{Character specifying which permutation scheme to apply for p-value calculation. When equal to \code{"all_genes"}, null distribution of p-values is calculated from all genes and the maximum number of permutation cycles is 10. When  \code{permutations = "per_gene"}, null distribution of p-values is calculated for each gene separately based on permutations of this individual gene. The latter approach may take a lot of computational time. We suggest using the first option.}
}
\value{
Returns a \code{\linkS4class{dmDStest}} or 
  \code{\linkS4class{dmSQTLtest}} object.
}
\description{
First, estimate the null Dirichlet-multinomial and beta-binomial model
parameters and likelihoods. Second, perform the gene-level (DM model) and
feature-level (BB model) likelihood ratio tests.
}
\details{
One must specify one of the arguments: \code{coef}, \code{design} or
  \code{contrast}.
}
\examples{

###################################
### Differential splicing analysis
###################################

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}}, 
  \code{\link{plotPValues}}, \code{\link{dmDispersion}}, \code{\link{dmFit}}
}

