% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdata.R, R/class_dmSQTLdata.R
\docType{methods}
\name{plotData}
\alias{plotData}
\alias{plotData,dmDSdata-method}
\alias{plotData,dmSQTLdata-method}
\title{Plot data summary}
\usage{
plotData(x, ...)

\S4method{plotData}{dmDSdata}(x)

\S4method{plotData}{dmSQTLdata}(x, plot_type = "features")
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdata}} or \code{\linkS4class{dmSQTLdata}} 
object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}

\item{plot_type}{Character specifying which type of histogram to plot. Possible
values \code{"features"}, \code{"snps"} or \code{"blocks"}.}
}
\value{
Returns a \code{ggplot} object and can be further modified, for
  example, using \code{theme()}. Plots a histogram of the number of features
  per gene. Additionally, for \code{\linkS4class{dmSQTLdata}} object, plots a
  histogram of the number of SNPs per gene and a histogram of the number of
  unique SNPs (blocks) per gene.
}
\description{
Plot data summary
}
\examples{
###################################
### Differential splicing analysis
###################################


#############################
### sQTL analysis
#############################

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}}, 
  \code{\link{plotDispersion}}, \code{\link{plotProportions}},
  \code{\link{plotPValues}}
}

