% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R, R/class_dmSQTLdispersion.R
\docType{methods}
\name{plotDispersion}
\alias{plotDispersion}
\alias{plotDispersion,dmDSdispersion-method}
\alias{plotDispersion,dmSQTLdispersion-method}
\title{Dispersion versus mean expression plot}
\usage{
plotDispersion(x, ...)

\S4method{plotDispersion}{dmDSdispersion}(x, out_dir = NULL)

\S4method{plotDispersion}{dmSQTLdispersion}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdispersion}} or
\code{\linkS4class{dmSQTLdispersion}} object.}

\item{...}{Other parameters that can be defined by methods using this
generic.}

\item{out_dir}{Character string that is used to save the plot
in \code{paste0(out_dir, plot_name, ".pdf")} file. \code{plot_name} depends
on type of a plot produced, for example, \code{plot_name = "hist_features"}
for histogram with number of features per gene. If \code{NULL},
the plot is returned as \code{ggplot} object and can be further modified,
for example, using \code{theme()}.}
}
\value{
Scatterplot of Dirichlet-multinomial gene-wise dispersion versus mean
gene expression. Both variables are scaled with log10. One dot in the plot
corresponds to a gene.
}
\description{
Dispersion versus mean expression plot
}
\examples{
###################################
### Differential splicing analysis
###################################
# If possible, use BPPARAM = BiocParallel::MulticoreParam() with more workers

d <- data_dmDSdata
\donttest{
### Filtering
# Check what is the minimal number of replicates per condition
table(samples(d)$group)
d <- dmFilter(d, min_samps_gene_expr = 7, min_samps_feature_expr = 3,
 min_samps_feature_prop = 0)

### Calculate dispersion
d <- dmDispersion(d, BPPARAM = BiocParallel::SerialParam())

plotDispersion(d)

}
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}},
  \code{\link{plotData}}, \code{\link{plotFit}}, \code{\link{plotTest}}
}

