% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{plotPValues}
\alias{plotPValues}
\alias{plotPValues,dmDStest-method}
\alias{plotPValues,dmSQTLtest-method}
\title{Plot p-values distribution}
\usage{
plotPValues(x, ...)

\S4method{plotPValues}{dmDStest}(x, level = "gene")

\S4method{plotPValues}{dmSQTLtest}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLtest}}
object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}

\item{level}{Character specifying which type of results to return. Possible
values \code{"gene"} or \code{"feature"}.}
}
\value{
Plot a histogram of p-values.
}
\description{
Plot p-values distribution
}
\examples{

###################################
### Differential splicing analysis
###################################

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}},
  \code{\link{plotData}}, \code{\link{plotDispersion}}, \code{\link{plotProportions}}
}

