% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{plotTest}
\alias{plotTest}
\alias{plotTest,dmDStest-method}
\alias{plotTest,dmSQTLtest-method}
\title{Plot p-values distribution}
\usage{
plotTest(x, ...)

\S4method{plotTest}{dmDStest}(x, out_dir = NULL)

\S4method{plotTest}{dmSQTLtest}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLtest}}
object.}

\item{...}{Other parameters that can be defined by methods using this
generic.}

\item{out_dir}{Character string that is used to save the plot 
in \code{paste0(out_dir, plot_name, ".pdf")} file. \code{plot_name} depends 
on type of a plot produced, for example, \code{plot_name = "hist_features"} 
for histogram with number of features per gene. If \code{NULL}, 
the plot is returned as \code{ggplot} object and can be further modified, 
for example, using \code{theme()}.}
}
\value{
Plot a histogram of p-values.
}
\description{
Plot p-values distribution
}
\examples{

###################################
### Differential splicing analysis
###################################
# If possible, use BPPARAM = BiocParallel::MulticoreParam() with more workers

d <- data_dmDSdata
\donttest{
### Filtering
# Check what is the minimal number of replicates per condition 
table(samples(d)$group)
d <- dmFilter(d, min_samps_gene_expr = 7, min_samps_feature_expr = 3, 
 min_samps_feature_prop = 0)

### Calculate dispersion
d <- dmDispersion(d, BPPARAM = BiocParallel::SerialParam())

### Fit full model proportions
d <- dmFit(d, BPPARAM = BiocParallel::SerialParam())

### Fit null model proportions and test for DS
d <- dmTest(d, BPPARAM = BiocParallel::SerialParam())

plotTest(d)


}
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}},
  \code{\link{plotData}}, \code{\link{plotDispersion}}, \code{\link{plotFit}}
}

