% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R, R/class_dmSQTLdispersion.R
\docType{methods}
\name{dmDispersion}
\alias{dmDispersion}
\alias{dmDispersion,dmDSdata-method}
\alias{dmDispersion,dmSQTLdata-method}
\title{Estimate dispersions in Dirichlet-multinomial model}
\usage{
dmDispersion(x, ...)

\S4method{dmDispersion}{dmDSdata}(x, design, mean_expression = TRUE,
  common_dispersion = TRUE, genewise_dispersion = TRUE,
  disp_adjust = TRUE, disp_subset = 0.1, disp_interval = c(0, 1e+05),
  disp_tol = 10, disp_init = 100, disp_grid_length = 21,
  disp_grid_range = c(-10, 10), disp_moderation = "trended",
  disp_prior_df = 0, disp_span = 0.1, one_way = TRUE,
  prop_mode = "constrOptim", prop_tol = 1e-12, coef_mode = "optim",
  coef_tol = 1e-12, verbose = 0,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))

\S4method{dmDispersion}{dmSQTLdata}(x, mean_expression = TRUE,
  common_dispersion = TRUE, genewise_dispersion = TRUE,
  disp_adjust = TRUE, disp_mode = "grid", disp_interval = c(0, 10000),
  disp_tol = 1e-08, disp_init = 100, disp_init_weirMoM = TRUE,
  disp_grid_length = 21, disp_grid_range = c(-10, 10),
  disp_moderation = "none", disp_prior_df = 0, disp_span = 0.1,
  prop_mode = "constrOptim", prop_tol = 1e-12, verbose = 0,
  speed = TRUE, BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdata}} or \code{\linkS4class{dmSQTLdata}}
object.}

\item{...}{Other parameters that can be defined by methods using this
generic.}

\item{design}{Numeric matrix definig the model that should be used when 
estimating dispersion. Normally this should be a full model design.}

\item{mean_expression}{Logical. Whether to estimate the mean expression of 
genes.}

\item{common_dispersion}{Logical. Whether to estimate the common dispersion.}

\item{genewise_dispersion}{Logical. Whether to estimate the gene-wise 
dispersion.}

\item{disp_adjust}{Logical. Whether to use the Cox-Reid adjusted or 
non-adjusted profile likelihood.}

\item{disp_subset}{Value from 0 to 1 defining the percentage of genes used in
common dispersion estimation. The default is 0.1, which uses 10% of 
randomly selected genes to speed up the dispersion estimation process. Use 
\code{set.seed} function to make the analysis reproducible. See Examples.}

\item{disp_interval}{Numeric vector of length 2 defining the interval of 
possible values for the common dispersion.}

\item{disp_tol}{The desired accuracy when estimating common dispersion.}

\item{disp_init}{Initial dispersion. If \code{common_dispersion} is 
\code{TRUE}, then \code{disp_init} is overwritten by common dispersion 
estimate.}

\item{disp_grid_length}{Length of the search grid.}

\item{disp_grid_range}{Vector giving the limits of grid interval.}

\item{disp_moderation}{Dispersion moderation method. One can choose to shrink
the dispersion estimates toward the common dispersion (\code{"common"}) or 
toward the (dispersion versus mean expression) trend (\code{"trended"})}

\item{disp_prior_df}{Degree of moderation (shrinkage) in case when it can not
be calculated automaticaly (number of genes on the upper boundary of grid 
is smaller than 10). By default it is equal to 0.}

\item{disp_span}{Value from 0 to 1 defining the percentage of genes used in 
smoothing sliding window when calculating the dispersion versus mean 
expression trend.}

\item{one_way}{Logical. Should the shortcut fitting be used when the design 
corresponds to multiple group comparison. If \code{TRUE} (the default), 
then proportions are fitted per group and regression coefficients are 
recalculated from these fittings.}

\item{prop_mode}{Optimization method used to estimate proportions. Possible 
value \code{"constrOptim"}.}

\item{prop_tol}{The desired accuracy when estimating proportions.}

\item{coef_mode}{Optimization method used to estimate regression
coefficients. Possible values \code{"optim"} (the default), \code{"nlminb"}
or \code{"Rcgmin"}.}

\item{coef_tol}{The desired accuracy when estimating regression coefficients.}

\item{verbose}{Numeric. Definie the level of progress messages displayed. 0 -
no messages, 1 - main messages, 2 - message for every gene fitting.}

\item{BPPARAM}{Parallelization method used by 
\code{\link[BiocParallel]{bplapply}}.}

\item{speed}{Logical. If \code{FALSE}, dispersion is calculated per each
gene-block. Such calculation may take a long time, since there can be
hundreds of SNPs/blocks per gene. If \code{TRUE}, there will be only one
dipsersion calculated per gene and it will be assigned to all the blocks
matched with this gene.}
}
\value{
Returns a \code{\linkS4class{dmDSdispersion}} or 
  \code{\linkS4class{dmSQTLdispersion}} object.
}
\description{
Maximum likelihood estimates of dispersion parameters in the
Dirichlet-multinomial model used in differential splicing or sQTL analysis.
}
\details{
Parameters that are used in the dispersion estimation start with 
  prefix \code{disp_}, and those that are used for the proportion estimation 
  start with \code{prop_}.
  
  There are 2 optimization methods implemented within dmDispersion 
  \code{"optimize"} for the common dispersion and \code{"grid"} for the 
  gene-wise dispersion.
  
  Arguments that are used by all the methods are:
  
  \itemize{ \item \code{disp_adjust} \item \code{prop_mode}: 
  \code{"constrOptim"} uses \code{\link{constrOptim}} function to maximize 
  the likelihood of Dirichlet-multinomial proportions. \item \code{prop_tol}:
  The accuracy for proportions estimation defined as \code{reltol} in 
  \code{\link{constrOptim}}. }
  
  Only some of the rest of dispersion parameters in dmDispersion have an 
  influence on a given optimization method. Here is a list of such active 
  parameters:
  
  \code{"optimize"}:
  
  \itemize{ \item \code{disp_interval}: Passed as \code{interval}. \item 
  \code{disp_tol}: The accuracy defined as \code{tol}. }
  
  \code{"grid"}, which uses the grid approach from \code{\link{edgeR}}:
  
  \itemize{ \item \code{disp_init}, \code{disp_grid_length}, 
  \code{disp_grid_range}: Parameters used to construct the search grid 
  \code{disp_init * 2^seq(from = disp_grid_range[1]}, \code{to = 
  disp_grid_range[2]}, \code{length = disp_grid_length)}. \item 
  \code{disp_moderation}: Dipsersion shrinkage is available only with 
  \code{"grid"} method. \item \code{disp_prior_df}: Used only when dispersion
  shrinkage is activated. Moderated likelihood is equal to \code{loglik + 
  disp_prior_df * moderation}. Higher \code{disp_prior_df}, more shrinkage 
  toward common or trended dispersion is applied. \item \code{disp_span}: 
  Used only when dispersion moderation toward trend is activated. }
}
\examples{
###################################
### Differential splicing analysis
###################################

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}}, 
  \code{\link{plotDispersion}}, \code{\link{dmFit}}, \code{\link{dmTest}}
}

