% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\docType{class}
\name{dmDSdata-class}
\alias{[,dmDSdata,ANY-method}
\alias{[,dmDSdata-method}
\alias{counts,dmDSdata-method}
\alias{dmDSdata-class}
\alias{length,dmDSdata-method}
\alias{names,dmDSdata-method}
\alias{samples}
\alias{samples,dmDSdata-method}
\title{dmDSdata object}
\usage{
\S4method{counts}{dmDSdata}(object)

samples(x, ...)

\S4method{samples}{dmDSdata}(x)

\S4method{names}{dmDSdata}(x)

\S4method{length}{dmDSdata}(x)

\S4method{[}{dmDSdata,ANY}(x, i, j)
}
\arguments{
\item{object, x}{dmDSdata object.}

\item{...}{Other parameters that can be defined by methods using 
this generic.}

\item{i, j}{Parameters used for subsetting.}
}
\value{
\itemize{
 \item \code{counts(object)}: Get a data frame with counts.
 \item \code{samples(x)}: Get a data frame with the sample information.
  \item \code{names(x)}: Get the gene names.
  \item \code{length(x)}: Get the number of genes.
  \item \code{x[i, j]}: Get a subset of dmDSdata object that consists of 
  counts for genes i and samples j. 
}
}
\description{
dmDSdata contains expression, in counts, of genomic features such as exons or 
transcripts and sample information needed for the differential splicing (DS) 
analysis. It can be created with function \code{\link{dmDSdata}}.
}
\section{Slots}{

\describe{
\item{\code{counts}}{\code{\linkS4class{MatrixList}} of expression, in counts, 
of genomic features. Rows correspond to genomic features, such as exons 
or transcripts. Columns correspond to samples. MatrixList is partitioned 
in a way that each of the matrices in a list contains counts for a single 
gene.}

\item{\code{samples}}{Data frame with information about samples. It must contain 
variables: \code{sample_id} of unique sample names and \code{group} which 
groups samples into conditions.}
}}
\examples{
###################################
### Differential splicing analysis
###################################

d <- data_dmDSdata

head(counts(d))
samples(d)
head(names(d))
length(d)
d[1:20, ]
d[1:20, 1:3]

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\linkS4class{dmDSdispersion}}, 
\code{\linkS4class{dmDSfit}}, \code{\linkS4class{dmDStest}}
}

