% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdata.R
\name{dmDSdata}
\alias{dmDSdata}
\title{Create dmDSdata object}
\usage{
dmDSdata(counts, gene_id, feature_id, sample_id, group)
}
\arguments{
\item{counts}{Numeric matrix or data frame of counts.
Rows represent features, for example, exons, exonic bins or transcripts.
Columns represent samples.}

\item{gene_id}{Vector of gene IDs corresponding to \code{counts}.}

\item{feature_id}{Vector of feature IDs corresponding to \code{counts}.}

\item{sample_id}{Vector of unique sample IDs corresponding to the columns
in \code{counts}.}

\item{group}{Vector that defines the grouping of samples.}
}
\value{
Returns a \linkS4class{dmDSdata} object.
}
\description{
Constructor function for a \code{\linkS4class{dmDSdata}} object.
}
\examples{
#############################
### Create dmDSdata object
#############################
### Get kallisto transcript counts from 'PasillaTranscriptExpr' package

library(PasillaTranscriptExpr)

data_dir  <- system.file("extdata", package = "PasillaTranscriptExpr")

metadata <- read.table(file.path(data_dir, "metadata.txt"), header = TRUE,
 as.is = TRUE)
metadata

counts <- read.table(file.path(data_dir, "counts.txt"), header = TRUE,
 as.is = TRUE)
head(counts)

# Create a dmDSdata object
d <- dmDSdata(counts = counts[, metadata$SampleName],
 gene_id = counts$gene_id, feature_id = counts$feature_id,
 sample_id = metadata$SampleName, group = metadata$condition)

plotData(d)

# Use a subset of genes, which is defined in the following file
gene_id_subset <- readLines(file.path(data_dir, "gene_id_subset.txt"))
d <- d[names(d) \%in\% gene_id_subset, ]

plotData(d)
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{plotData}}, \code{\link{dmFilter}},
\code{\link{dmDispersion}}, \code{\link{dmFit}}, \code{\link{dmTest}}
}

