% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R
\docType{class}
\name{dmDSdispersion-class}
\alias{common_dispersion}
\alias{common_dispersion,dmDSdispersion-method}
\alias{common_dispersion<-}
\alias{common_dispersion<-,dmDSdispersion-method}
\alias{dmDSdispersion-class}
\alias{genewise_dispersion}
\alias{genewise_dispersion,dmDSdispersion-method}
\alias{genewise_dispersion<-}
\alias{genewise_dispersion<-,dmDSdispersion-method}
\alias{mean_expression}
\alias{mean_expression,dmDSdispersion-method}
\title{dmDSdispersion object}
\usage{
mean_expression(x, ...)

\S4method{mean_expression}{dmDSdispersion}(x)

common_dispersion(x, ...)

\S4method{common_dispersion}{dmDSdispersion}(x)

common_dispersion(x) <- value

\S4method{common_dispersion}{dmDSdispersion}(x) <- value

genewise_dispersion(x, ...)

\S4method{genewise_dispersion}{dmDSdispersion}(x)

genewise_dispersion(x) <- value

\S4method{genewise_dispersion}{dmDSdispersion}(x) <- value
}
\arguments{
\item{x}{dmDSdispersion object.}

\item{...}{Other parameters that can be defined by methods using this
  generic.}

\item{value}{Values that replace current attributes.}
}
\value{
\itemize{ \item \code{mean_expression(x)}: Get a data frame with mean gene
expression. \item \code{common_dispersion(x), common_dispersion(x) <- value}:
Get or set common dispersion. \code{value} must be numeric. \item
\code{genewise_dispersion(x), genewise_dispersion(x) <- value}: Get a data
frame with gene-wise dispersion or set new gene-wise dispersion. \code{value}
must be a vector. }
}
\description{
dmDSdispersion extends the \code{\linkS4class{dmDSdata}} by adding the
dispersion estimates of Dirichlet-multinomial distribution used to model the
feature (e.g., transcript, exon, exonic bin) coiunts for each gene in the
differential splicing analysis. Result of \code{\link{dmDispersion}}.
}
\section{Slots}{

\describe{
\item{\code{mean_expression}}{Numeric vector of mean gene expression.}

\item{\code{common_dispersion}}{Numeric value of estimated common dispersion.}

\item{\code{genewise_dispersion}}{Numeric vector of estimated gene-wise dispersions.}
}}
\examples{
###################################
### Differential splicing analysis
###################################
# If possible, use BPPARAM = BiocParallel::MulticoreParam() with more workers

d <- data_dmDSdata
\donttest{
### Filtering
# Check what is the minimal number of replicates per condition
table(samples(d)$group)
d <- dmFilter(d, min_samps_gene_expr = 7, min_samps_feature_expr = 3,
 min_samps_feature_prop = 0)

### Calculate dispersion
d <- dmDispersion(d, BPPARAM = BiocParallel::SerialParam())
plotDispersion(d)

head(mean_expression(d))
common_dispersion(d)
head(genewise_dispersion(d))

}
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\linkS4class{dmDSdata}},
  \code{\linkS4class{dmDSfit}}, \code{\linkS4class{dmDStest}}
}

