% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSfit.R
\docType{class}
\name{dmDSfit-class}
\alias{coefficients,dmDSfit-method}
\alias{dmDSfit-class}
\alias{proportions}
\alias{proportions,dmDSfit-method}
\title{dmDSfit object}
\usage{
proportions(x, ...)

\S4method{proportions}{dmDSfit}(x)

\S4method{coefficients}{dmDSfit}(object, level = "gene")
}
\arguments{
\item{x}{dmDSdispersion object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}

\item{level}{Character specifying which type of results to return. Possible
values \code{"gene"} or \code{"feature"}.}
}
\value{
\itemize{ \item \code{proportions(x)}: Get a data frame with estimated 
feature ratios for each sample. \item \code{coefficients(x)}:  }
}
\description{
dmDSfit extends the \code{\linkS4class{dmDSdispersion}} class by adding the 
full model Dirichlet-multinomial feature proportion estimates needed for the 
differential splicing analysis. Feature ratios are estimated for each gene 
and each condition. Result of \code{\link{dmFit}}.
}
\section{Slots}{

\describe{
\item{\code{design_fit_full}}{Numeric matrix of the desing used to fit the full 
model.}

\item{\code{fit_full}}{\code{\linkS4class{MatrixList}} containing estimated feature 
ratios in each sample based on the full Dirichlet-multinomial (DM) model.}

\item{\code{lik_full}}{Numeric vector of the per gene DM full model likelihoods.}

\item{\code{coef_full}}{\code{\linkS4class{MatrixList}} with the regression 
coefficients based on the DM model}

\item{\code{fit_full_bb}}{\code{\linkS4class{MatrixList}} containing estimated
feature ratios in each sample based on the full beta-binomial (BB) model.}

\item{\code{lik_full_bb}}{Numeric vector of the per gene BB full model likelihoods.}

\item{\code{coef_full_bb}}{\code{\linkS4class{MatrixList}} with the regression 
coefficients based on the BB model}
}}
\examples{

###################################
### Differential splicing analysis
###################################

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\linkS4class{dmDSdata}}, 
  \code{\linkS4class{dmDSdispersion}}, \code{\linkS4class{dmDStest}}
}

