% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmSQTLtest.R
\docType{class}
\name{dmSQTLtest-class}
\alias{dmSQTLtest-class}
\alias{results,dmSQTLtest-method}
\title{dmSQTLtest object}
\usage{
\S4method{results}{dmSQTLtest}(x)
}
\arguments{
\item{x}{dmSQTLtest object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}
}
\value{
\itemize{ \item \code{results(x)}: Get a data frame with results. See Slots. 
}
}
\description{
dmSQTLtest extends the \code{\linkS4class{dmSQTLfit}} class by adding the 
null model Dirichlet-multinomial feature proportion estimates and the results
of testing for sQTLs. Proportions are calculated for each gene-block pair 
from pooled (no grouping into conditions) counts. Result of 
\code{\link{dmTest}}.
}
\section{Slots}{

\describe{
\item{\code{lik_null}}{List of numeric vectors with the per gene-snp DM null model
likelihoods.}

\item{\code{results_gene}}{Data frame with the gene-level results including: 
\code{gene_id} - gene IDs, \code{block_id} - block IDs, \code{snp_id} - SNP
IDs, \code{lr} - likelihood ratio statistics based on the DM model, 
\code{df} - degrees of freedom, \code{pvalue} - p-values estimated based on
permutations and \code{adj_pvalue} - Benjamini & Hochberg adjusted 
p-values.}
}}

\examples{

#############################
### sQTL analysis
#############################
}
\seealso{
\code{\link{data_dmSQTLdata}}, \code{\linkS4class{dmSQTLdata}}, 
  \code{\linkS4class{dmSQTLdispersion}}, \code{\linkS4class{dmSQTLfit}}
}
\author{
Malgorzata Nowicka
}
