% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmSQTLtest.R
\docType{class}
\name{dmSQTLtest-class}
\alias{dmSQTLtest-class}
\alias{results,dmSQTLtest-method}
\title{dmSQTLtest object}
\usage{
\S4method{results}{dmSQTLtest}(x)
}
\arguments{
\item{x}{dmSQTLtest object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}
}
\value{
\itemize{ \item \code{results(x)}: Get a data frame with results. See Slots. 
}
}
\description{
dmSQTLtest extends the \code{\linkS4class{dmSQTLfit}} class by adding the 
null model Dirichlet-multinomial feature proportion estimates and the results
of testing for sQTLs. Proportions are calculated for each gene-block pair 
from pooled (no grouping into conditions) counts. Result of 
\code{\link{dmTest}}.
}
\section{Slots}{

\describe{
\item{\code{fit_null}}{List of \code{\linkS4class{MatrixList}}. Each of them contains
null proportions, likelihoods and degrees of freedom for all the blocks 
(unique SNPs) assigned to a given gene.}

\item{\code{results}}{Data frame with \code{gene_id} - gene IDs, \code{block_id} - 
block IDs, \code{snp_id} - SNP IDs, \code{lr} - likelihood ratio 
statistics, \code{df} - degrees of freedom, \code{pvalue} - p-values
estimated based on permutations and \code{adj_pvalue} - Benjamini &
Hochberg adjusted p-values.}
}}
\examples{

#############################
### sQTL analysis
#############################
# If possible, use BPPARAM = BiocParallel::MulticoreParam() with more workers

d <- data_dmSQTLdata
\donttest{
### Filtering
d <- dmFilter(d, min_samps_gene_expr = 70, min_samps_feature_expr = 5, 
   min_samps_feature_prop = 0, minor_allele_freq = 5, 
   BPPARAM = BiocParallel::SerialParam())

### Calculate dispersion
d <- dmDispersion(d, BPPARAM = BiocParallel::SerialParam())

### Fit full model proportions
d <- dmFit(d, BPPARAM = BiocParallel::SerialParam())

### Fit null model proportions and test for sQTLs
d <- dmTest(d, BPPARAM = BiocParallel::SerialParam())
plotPValues(d)

head(results(d))

}
}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmSQTLdata}}, \code{\linkS4class{dmSQTLdata}}, 
  \code{\linkS4class{dmSQTLdispersion}}, \code{\linkS4class{dmSQTLfit}}
}

