% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDStest.R, R/class_dmSQTLtest.R
\docType{methods}
\name{dmTest}
\alias{dmTest}
\alias{dmTest,dmDSfit-method}
\alias{dmTest,dmSQTLfit-method}
\title{Likelihood ratio test}
\usage{
dmTest(x, ...)

\S4method{dmTest}{dmDSfit}(x, compared_groups = levels(samples(x)$group),
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = 0,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))

\S4method{dmTest}{dmSQTLfit}(x, permutations = "all_genes",
  prop_mode = "constrOptimG", prop_tol = 1e-12, verbose = 0,
  BPPARAM = BiocParallel::MulticoreParam(workers = 1))
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}} or \code{\linkS4class{dmSQTLfit}}
object.}

\item{...}{Other parameters that can be defined by methods using this
generic.}

\item{compared_groups}{Vector that defines which experimental conditions
should be tested for differential splicing. By default, we test for a
difference between any of the groups specified in \code{samples(x)$group}.
Values in this vector should indicate levels or numbers of levels in
\code{samples(x)$group}.}

\item{prop_mode}{Optimization method used to estimate proportions. Possible 
values \code{"constrOptim"} and \code{"constrOptimG"}.}

\item{prop_tol}{The desired accuracy when estimating proportions.}

\item{verbose}{Numeric. Definie the level of progress messages displayed. 0 -
no messages, 1 - main messages, 2 - message for every gene fitting.}

\item{BPPARAM}{Parallelization method used by 
\code{\link[BiocParallel]{bplapply}}.}

\item{permutations}{Character specifying which permutation scheme to apply for p-value calculation. When equal to \code{"all_genes"}, null distribution of p-values is calculated from all genes and the maximum number of permutation cycles is 10. When  \code{permutations = "per_gene"}, null distribution of p-values is calculated for each gene separately based on permutations of this individual gene. The latter approach may take a lot of computational time. We suggest using the first option.}
}
\value{
Returns a \code{\linkS4class{dmDStest}} or
  \code{\linkS4class{dmSQTLtest}} object.
}
\description{
First, estimate the null Dirichlet-multinomial model proportions, i.e.,
feature ratios are estimated based on pooled (no grouping into conditions)
counts. Use the likelihood ratio statistic to test for the difference between
feature proportions in different groups to identify the differentially
spliced genes (differential splicing analysis) or the sQTLs (sQTL analysis).
}
\examples{

###################################
### Differential splicing analysis
###################################
# If possible, use BPPARAM = BiocParallel::MulticoreParam() with more workers

d <- data_dmDSdata
\donttest{
### Filtering
# Check what is the minimal number of replicates per condition 
table(samples(d)$group)
d <- dmFilter(d, min_samps_gene_expr = 7, min_samps_feature_expr = 3, 
 min_samps_feature_prop = 0)

### Calculate dispersion
d <- dmDispersion(d, BPPARAM = BiocParallel::SerialParam())

### Fit full model proportions
d <- dmFit(d, BPPARAM = BiocParallel::SerialParam())

### Fit null model proportions and test for DS
d <- dmTest(d, BPPARAM = BiocParallel::SerialParam())
plotTest(d)

head(proportions(d))
head(statistics(d))
head(results(d))

}

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}},
  \code{\link{plotTest}}, \code{\link{dmDispersion}}, \code{\link{dmFit}}
}

