% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDStest.R
\docType{methods}
\name{dmTwoStageTest}
\alias{dmTwoStageTest}
\alias{dmTwoStageTest,dmDStest-method}
\title{Two-stage test}
\usage{
dmTwoStageTest(x, ...)

\S4method{dmTwoStageTest}{dmDStest}(x, FDR = 0.05, verbose = 0,
  BPPARAM = BiocParallel::SerialParam())
}
\arguments{
\item{x}{\code{\linkS4class{dmDStest}} or \code{\linkS4class{dmSQTLtest}}
object.}

\item{...}{Other parameters that can be defined by methods using this
generic.}

\item{FDR}{Numeric. Cutoff for the FDR.}

\item{verbose}{Numeric. Definie the level of progress messages displayed. 0 -
no messages, 1 - main messages, 2 - message for every gene fitting.}

\item{BPPARAM}{Parallelization method used by 
\code{\link[BiocParallel]{bplapply}}.}
}
\value{
Returns a data frame with adjusted feature-level p-values.
}
\description{
Two-stage test
}

