% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdata.R, R/class_dmSQTLdata.R
\docType{methods}
\name{plotData}
\alias{plotData}
\alias{plotData,dmDSdata-method}
\alias{plotData,dmSQTLdata-method}
\title{Plot data summary}
\usage{
plotData(x, ...)

\S4method{plotData}{dmDSdata}(x, out_dir = NULL)

\S4method{plotData}{dmSQTLdata}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdata}} or \code{\linkS4class{dmSQTLdata}}
object.}

\item{...}{Other parameters that can be defined by methods using this
generic.}

\item{out_dir}{Character string that is used to save the plot 
in \code{paste0(out_dir, plot_name, ".pdf")} file. \code{plot_name} depends 
on type of a plot produced, for example, \code{plot_name = "hist_features"} 
for histogram with number of features per gene. If \code{NULL}, 
the plot is returned as \code{ggplot} object and can be further modified, 
for example, using \code{theme()}.}
}
\value{
Plot a histogram of the number of features per gene. Additionally, 
for \code{\linkS4class{dmSQTLdata}} object, plot a histogram of the number of
SNPs per gene and a histogram of the number of unique SNPs (blocks) per gene.
}
\description{
Plot data summary
}
\examples{
###################################
### Differential splicing analysis
###################################

d <- data_dmDSdata
plotData(d)



#############################
### sQTL analysis
#############################

d <- data_dmSQTLdata
plotData(d)

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}}, 
\code{\link{plotDispersion}}, \code{\link{plotFit}}, \code{\link{plotTest}}
}

