% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R,
%   R/class_dmSQTLdispersion.R
\docType{methods}
\name{plotDispersion}
\alias{plotDispersion}
\alias{plotDispersion,dmDSdispersion-method}
\alias{plotDispersion,dmSQTLdispersion-method}
\title{Precision versus mean expression plot}
\usage{
plotDispersion(x, ...)

\S4method{plotDispersion}{dmDSdispersion}(x)

\S4method{plotDispersion}{dmSQTLdispersion}(x)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdispersion}} or 
\code{\linkS4class{dmSQTLdispersion}} object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}
}
\value{
Normally in the differential analysis based on RNA-seq data, such 
  plot has dispersion parameter plotted on the y-axis. Here, the y-axis 
  represents precision since in the Dirichlet-multinomial model this is the 
  parameter that is directly estimated. It is important to keep in mind that
  the precision parameter (gamma0) is inverse proportional to dispersion 
  (theta): theta = 1 / (1 + gamma0). In RNA-seq data, we can typically
  observe a trend where the dispersion decreases (here, presicion increases)
  for genes with higher mean expression.
}
\description{
Precision versus mean expression plot
}
\examples{
# --------------------------------------------------------------------------
# Create dmDSdata object 
# --------------------------------------------------------------------------
## Get kallisto transcript counts from the 'PasillaTranscriptExpr' package

library(PasillaTranscriptExpr)
\donttest{
data_dir  <- system.file("extdata", package = "PasillaTranscriptExpr")

## Load metadata
metadata <- read.table(file.path(data_dir, "metadata.txt"), header = TRUE, 
  as.is = TRUE)

## Load counts
counts <- read.table(file.path(data_dir, "counts.txt"), header = TRUE, 
  as.is = TRUE)

## Create a samples data frame
samples <- data.frame(sample_id = metadata$SampleName, 
  group = metadata$condition)
levels(samples$group)

## Create a dmDSdata object
d <- dmDSdata(counts = counts, samples = samples)

## Use a subset of genes, which is defined in the following file
gene_id_subset <- readLines(file.path(data_dir, "gene_id_subset.txt"))

d <- d[names(d) \%in\% gene_id_subset, ]

# --------------------------------------------------------------------------
# Differential transcript usage analysis - simple two group comparison 
# --------------------------------------------------------------------------

## Filtering
## Check what is the minimal number of replicates per condition 
table(samples(d)$group)

d <- dmFilter(d, min_samps_gene_expr = 7, min_samps_feature_expr = 3,
  min_gene_expr = 10, min_feature_expr = 10)

plotData(d)

## Create the design matrix
design <- model.matrix(~ group, data = samples(d))

## To make the analysis reproducible
set.seed(123)
## Calculate dispersion
d <- dmDispersion(d, design = design)

plotDispersion(d)

head(mean_expression(d))
common_dispersion(d)
head(genewise_dispersion(d))
}
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}},
  \code{\link{plotData}}, \code{\link{plotProportions}}, \code{\link{plotPValues}}
}
\author{
Malgorzata Nowicka
}
