% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSdispersion.R, R/class_dmSQTLdispersion.R
\docType{methods}
\name{plotDispersion}
\alias{plotDispersion}
\alias{plotDispersion,dmDSdispersion-method}
\alias{plotDispersion,dmSQTLdispersion-method}
\title{Dispersion versus mean expression plot}
\usage{
plotDispersion(x, ...)

\S4method{plotDispersion}{dmDSdispersion}(x)

\S4method{plotDispersion}{dmSQTLdispersion}(x, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSdispersion}} or 
\code{\linkS4class{dmSQTLdispersion}} object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}
}
\value{
Normally in the differential analysis based on RNA-seq data, such 
  plot has dispersion parameter plotted on the y-axis. Here, the y-axis 
  represents precision since in the Dirichlet-multinomial model this is a 
  parameter that is  directly estimated. It is important to keep in mind that
  the precision parameter (gamma0) is inverse proportional to dispersion 
  (theta): theta = 1 / (1 + gamma0). In RNA-seq data, we can typically
  observe a trend where the dispersion decreases (here, presicion increases)
  for genes with higher mean expression.
}
\description{
Dispersion versus mean expression plot
}
\examples{
###################################
### Differential splicing analysis
###################################

}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}},
  \code{\link{plotData}}, \code{\link{plotProportions}}, \code{\link{plotPValues}}
}

