% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_dmDSfit.R, R/class_dmSQTLfit.R
\docType{methods}
\name{plotProportions}
\alias{plotProportions}
\alias{plotProportions,dmDSfit-method}
\alias{plotProportions,dmSQTLfit-method}
\title{Plot feature proportions}
\usage{
plotProportions(x, ...)

\S4method{plotProportions}{dmDSfit}(x, gene_id, group_variable,
  plot_type = "barplot", order = TRUE, plot_fit = TRUE,
  plot_main = TRUE, group_colors = NULL, feature_colors = NULL)

\S4method{plotProportions}{dmSQTLfit}(x, gene_id, snp_id,
  plot_type = "boxplot1", order = TRUE, plot_full = TRUE,
  plot_main = TRUE, out_dir = NULL)
}
\arguments{
\item{x}{\code{\linkS4class{dmDSfit}}, \code{\linkS4class{dmDStest}} or 
\code{\linkS4class{dmSQTLfit}}, \code{\linkS4class{dmSQTLtest}} object.}

\item{...}{Other parameters that can be defined by methods using this 
generic.}

\item{gene_id}{Character indicating a gene ID to be plotted.}

\item{group_variable}{Character indicating the grouping variable which is one
of the columns in the \code{samples} slot of \code{x}.}

\item{plot_type}{Character defining the type of the plot produced. Possible 
values \code{"barplot"}, \code{"boxplot1"}, \code{"boxplot2"}, 
\code{"lineplot"}, \code{"ribbonplot"}.}

\item{order}{Logical. Whether to plot the features ordered by their 
expression.}

\item{plot_main}{Logical. Whether to plot a title with the information about 
the Dirichlet-multinomial estimates.}

\item{group_colors}{Character vector with colors for each group defined by 
\code{group_variable}.}

\item{feature_colors}{Character vector with colors for each feature of gene
defined by \code{gene_id}.}

\item{snp_id}{Character indicating a SNP ID to be plotted. \code{snp_id} must
match \code{gene_id}.}

\item{plot_full}{Logical. Whether to plot the proportions estimated by the 
full model.}
}
\value{
Plot, per gene, the observed and estimated with Dirichlet-multinomial
  model feature proportions. Estimated group proportions are marked with
  diamond shapes.
}
\description{
This plot is available only for a group design, i.e., a comparison between 
groups.
}
\examples{

###################################
### Differential splicing analysis
###################################


}
\author{
Malgorzata Nowicka
}
\seealso{
\code{\link{data_dmDSdata}}, \code{\link{data_dmSQTLdata}}, 
  \code{\link{plotData}}, \code{\link{plotDispersion}}, 
  \code{\link{plotPValues}}
}

